+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Fix bazel key padding for short keys.
 - Fix PUT error path race.
+- Fix accept error busy loop.
 
 ## [0.7] - 2026-05-18